requirements.txt
streamlit==1.18.1
Pillow==9.3.0
opencv-python==4.6.0.66
numpy==1.23.5
scipy==1.10.0
huffman==0.2.0
pywavelets==1.4.1
pydub==0.25.1
moviepy==1.0.3

